+++ static/js/src/entry/account/form/components/TextInput.vue
@@ -42,7 +42,7 @@
 </template>
 
 <script>
-import formElementMixin from '../mixins/form-element';
+import formElementMixin from '../mixins/element';
 
 export default {
   name: 'TextInput',
