+++ static/js/src/entry/account/routes/blast-payments/Index.vue
@@ -23,7 +23,7 @@
 <script>
 /* eslint-disable camelcase */
 
-import routeMixin from '../../mixins/route';
+import routeMixin from '../mixin';
 import userMixin from '../../store/user/mixin';
 import Help from '../home/components/Help.vue';
 import RouteLoader from '../home/components/RouteLoader.vue';
