+++ static/js/src/entry/account/routes/payments/Index.vue
@@ -31,7 +31,7 @@
 <script>
 /* eslint-disable camelcase */
 
-import routeMixin from '../../mixins/route';
+import routeMixin from '../mixin';
 import userMixin from '../../store/user/mixin';
 import RouteLoader from '../home/components/RouteLoader.vue';
 import Appeal from '../home/containers/AppealContainer.vue';
